Add page routes for courses and about
