+++ lib/widgets/navBar/navbar_item.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class NavBarItem extends StatelessWidget {
+  final String title;
+  const NavBarItem(this.title, {super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Text(
+      title,
+      style: const TextStyle(
+        fontSize: 18,
+      ),
+    );
+  }
+}
