+++ lib/widgets/navBar/navbar_logo.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class NavBarLogo extends StatelessWidget {
+  const NavBarLogo({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: const SizedBox(
+        height: 80,
+        width: 150,
+        child: Image(
+          image: AssetImage('assets/logo.png'),
+        ),
+      ),
+    );
+  }
+}
