+++ lib/widgets/navBar/navigation_bar_desktop.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+
+import 'navbar_item.dart';
+import 'navbar_logo.dart';
+
+class NavigationBarDesktop extends StatelessWidget {
+  const NavigationBarDesktop({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 100,
+      child: const Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          NavBarLogo(),
+          Row(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              NavBarItem("Courses"),
+              SizedBox(
+                width: 60,
+              ),
+              NavBarItem("About")
+            ],
+          )
+        ],
+      ),
+    );
+  }
+}
