+++ lib/widgets/navBar/navigation_bar_mobile.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+import 'package:psa_codesprint/widgets/navbar_logo.dart';
+
+class NavigationBarMobile extends StatelessWidget {
+  const NavigationBarMobile({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 80,
+      child: Row(
+        mainAxisSize: MainAxisSize.max,
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          IconButton(
+              onPressed: () {},
+              icon: Icon(Icons.menu)),
+          NavBarLogo()
+        ],
+      ),
+    );
+  }
+}
